Effective Path Module
To find an effective path between all the selected cities using the Travelling Salesman Algorithm. Genetic Algorithm is a search technique used in computing to find the optimal solution to a computational problem that maximizes or minimizes a particular function. Genetic Algorithm is used to solve the Travelling Salesman Problem where one has to find the shortest route among the cities from the origin. The Travelling Salesman Problem (TSP) is well known in the field of combinatorial optimization.